Finished Auto Switcher Implemented
A working version of the auto switcher has been finished. However, there's a few things to keep in mind. The definitions for the commands are all commented out, since they stopped working ever since we switched to the Totes Drive. Totes Drive has it's own autonomous system in the Drivebase subsystem which the auto switcher has not been configured to work with; (instead it's still working with the same system we had before for auto). For now, the line for running the selected auto chooser command has been commented out and instead is still running the selected auto inside the Totes Drivebase.